advanced-api-project/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-filter==25.1